Run new query planner off of Tokio threads
Planning code is synchronous (it never `await`s)
and can take non-trivial time, so it is effectively "blocking"
from Tokio’s point of view. If too many Tokio threads are busy
planning queries, few threads remain available to run async code.

This runs the planner inside `tokio::task::spawn_blocking`,
following the precedent of graphql parsing in `query_analysis.rs`.
This uses a thread pool of maximum size 512 threads,
which may not be what we want.
From its documentation, `spawn_blocking` appears to be made
for code that often blocks in synchronous syscalls.

Instead we may prefer to have a thread pool for "compute tasks"
that defaults to $NUM_CPUS threads, and wrap it with something
that uses `tokio::sync::oneshot` to send results back to async code.