Require (...) not $(...) for PathList::Expr when already parsing LitExpr.
This refinement effectively means $(...) is necessary only when you want
to switch from normal JSONSelection parsing to LitExpr parsing. Once
you're already parsing a LitExpr, (...) works (without the $) for
wrapping nested literal values so you can apply a .key or ->method to
them. If you don't need to apply a .key or ->method to the literal, you
don't need parentheses at all.

We could potentially also use this PathParsingMode information to make
the leading '.' character optional for single-key .key paths (since
there's no ambiguity with field selections within a LitExpr), but I've
kept that parsing behavior unchanged for now.

I considered allowing either $(...) or (...) (or no parens) when parsing
a LitExpr, but that's not a semantically meaningful distinction, so I'd
rather force the syntax to be mode-specific, so different programmers
don't accidentally adopt different conventions.