Fix APQ and introspection gauges
Similar to the query planner gauges, these must be initialized after telemetry has been activated.

Fixes:
* `apollo_router_cache_size` for `APQ` and `introspection`

As a result of this change every query planner in the query planner pool will share the same cache storage for introspection queries. Otherwise many identical gauges are created, and they are not summed. It doesn't really make sense that each query planner should have a separate cache anyway.

Another issue was noticed, and commented on. APQLayer uses deduplicating cache which is fallible. This is initialized after telemetry has been activated, so if it fails it is likely that telemetry will stop working. This is not fixable without a larger refactor to bring APQ init forward, which is out of scope for this PR. We'll tackle router lifecycle as a larger holistic refactor later.