Only run fuzz on the `test` job when `fuzz` is requested and skip on releases.
The `fuzz` job uses the nightly version of Rust, necessarily.  However, due
to the predictability of nightly releases, they can start to fail from time
to time.  This is particularly problematic on release jobs where it is
necessary for the jobs to all pass in order to unblock the release.

Therefore, this makes the `fuzz` job disabled by default and only runs it
when requested and only requests it on non-release jobs (effectively, that
narrows it down to `ci_checks`).