Temporarily ping CI to Rust nightly-2024-09-22
We use Rust nightly on CI in order to check that fuzz targets can still compile.
With current latest nightly we’re running into a compiler panic:
https://github.com/rust-lang/rust/issues/130769

To unblock merging PRs, this pins the nightly version used on CI
to a slightly older one that does not have this panic.
This should be reverted after the panic is fixed in a new Rust nightly.