Allow `NamedSelection ::= Path SubSelection` without an `Alias`.
As long as a `Path` is followed by a `SubSelection`, we can merge its
(statically known) output properties together with other named
selections within a larger `SubSelection`, allowing much more succinct
extraction of similarly nested properties:

    selection: """
    id
    created
    model

    # New syntax:
    choices->first.message { role content }
    """

This selection generates an output object with the properties `id`,
`created`, `model`, `role`, and `content`, where the `role` and
`content` properties are taken from the `choices->first.message` object.

This new syntax respects guiding principle #3 (safe subsetting), because
a selection consisting of only

    choices->first.message { role content }

is equivalent to the selection

    role: choices->first.message.role
    content: choices->first.message.content

and continues to contribute the `role` and `content` properties in the
same way even when you add/remove other named sibling properties to/from
the selection (such as `id`, `created`, and `model`).

Thanks to @dylan-apollo for brainstorming the best syntax for achieving
this ergonomic improvement.