Inline InlineFragmentData::can_rebase_on
This has the same name as InlineFragment::can_rebase_on, which will be a
problem when I merge the two structures.

It's only used in one place: computing the parent type of the inline
fragment's subselection when it is being rebased. I inlined the function
and I think it's more readable this way as well.