Fix introspection false mismatch: defaultValue list of objects
In GraphQL introspection response, the `defaultValue` of an anrgument
or input object field is represented as a JSON string that contains
GraphQL syntax. If that value contains a GraphQL object value,
graphql-js seems to sort its contents by input object field name
whereas apollo-compiler preserves the source ordering.

This causes introspection JSON responses to be different in a way
that we consider not to be meaningful: object values are semantically
unordered maps.
Introspection "both" mode already had code to sort object values in order
to ignore these differences, but was missing the case where `defaultValue`
is not directly an object value but a list value of object values.