Forbid single-`.key` paths, requiring `$.key` instead.
Now that PR #6076 allows `PathWithSubSelection` syntax to appear within
a `SubSelection` alongside other named selections, the shorthand of
using `.field` to refer to a path consisting of a single `field` key has
become potentially ambiguous:

```graphql
__typename: $("Product")
.field { nested child }
```

Because parsing is insensitive to spaces/newlines/comments between
tokens, this code parses the same as

```graphql
__typename: $("Product").field { nested child }
```

which produces a runtime `ApplyToError` because the `"Product"` string
does not have a `field` property.

Fortunately, using the `$` variable (i.e. `$.field`) has identical
behavior to the shorthand, and does not suffer from any parsing
ambiguity:

```graphql
__typename: $("Product")
$.field { nested child }
```

This commit statically forbids using the `.field` shorthand, so
`$.field` is effectively required, which is a breaking change, but it's
better to implement this restriction while we're still in preview than
after we've publicly released this code.

Subjectively, I find `$.field` more explicit and harder to misread than
`.field`, so I believe this restriction is an improvement.