fix(federation): fix left-behind selections in fragment reuse
Sometimes, fragment reuse would leave behind a selection in a selection
set that is also part of a fragment that was applied to that selection
set. That's suboptimal.

An example from the apollo codebase is this:
```graphql
descendants {
  ...TraceNodeFragment
  errors {
    locations {
      __typename
      line
      column
    }
  }
}
fragment TraceNodeFragment on Trace {
  errors {
    locations {
      __typename
      line
      column
    }
  }
}
```

The cause is this difference in our `intersection` implementation:
https://github.com/apollographql/federation/blob/df5eb3cb0e2b4802fcd425ab9c23714de2707db3/internals-js/src/operations.ts#L2676-L2680

If a field selection's subselection completely disappears in
`intersection()`, we previously kept the field selection, instead of
removing it entirely.

By returning `None` we will now correctly remove the duplicate `errors` selection.

I did not add a unit test for this as we do not expect to rely on
fragment reuse long-term.