fix(federation): match JS condition order with multiple skip/include on the same selection
Rust was iterating in a consistent order, JS iterates in declaration
order.

Surprisingly this was the only place we use the consistent order. This
removes the `iter_sorted()` method from DirectiveList that is now no
longer used. Internally, DirectiveList still uses the consistent sort
order for hashing and equality.

I think it's somewhat nicer to use the consistent order here, so you do
not get different plans based on an innocuous input difference, but so
be it. We can consider using a consistent order everywhere we handle
directives in the future.