Enough functionality to implement adapative load shedding
An investigation into backpressure issues in the router.

Most of the changes are in various plugins to implement backpressure.
However, those fixes are not enough to provide useful functionality...

The current implementation of the router create a new pipeline for each
connection. This has the unfortunate impact of discarding state which is
required for various load impacting layers to work correctly.

This exploration modifies the router to hold a single master pipeline
which is clone'd for each connection. This allows the various tower
connection limiting layers to work correctly.

I've got a version which works with standard tower layers, commented out
here, but I've also got a potentially more interesting version which
uses a load shedded based on Little's Law, which is what is active in
this code.

Notes:

Modifying the pipeline to be cloneable has generally worked fine, but it
has caused issues for the Limit layer. This layer looks "generally
problematic" since it appears to make a number of assumptions about what
request rejection actually means. I've done some minimal modification to
try and make it work win a cloned pipeline, but tests are still failing
and I'm not sure it does what it should do.

I also noticed that when implementing backpressure, various mock tests
needed to be modified since test rejection happened earlier in the
pipeline and a map_result() somewhere isn't triggered. That needs some
investigation, but I think it's a small problem to address.

I modified the bridge query planner pool to prevent excessive queueing
in this layer. Since I now want to control this by load_shedding before
this service is reached, I only want enough channels to support the
number of planners.

Summary:

This PR provides a router which will operate with approximately
the same performance of the base router, but which controls memory and
rejects excess load to prevent "over-commit" by the router. This is a
very desirable property.

More testing is required, but this is looking promising so far.