If subgraph batching, do not log the full response when failing to notify a waiter
If a subgraph batch succeeds, we try to notify the original waiting
requester that the operation succeeded. If this notification fails, we
currently log out the entire subgraph response as part of the
notification error.

This may contain either or both of a lot of data or PII data which
should not be logged.