Fix header propagation where renames have already taken place
#4535 introduced an issue where the following config would not work:

```yaml
headers:
- propagate:
    named: a
    rename: b
- propagate:
    named: a
    rename: c
```

The existing logic maintained a set of headers that had been propagated already, but did not take into account renaming. It used the original name of the header when tracking which headers had been propagated.

This fix adds a test for propagation and fixes the logic to take into account renames.