pin point auth tests that fail comparison
These three examples result in mismatches when comparing PlanNodes between JS and Rust query plans. This is because the JS PlanNode is the one that will be used when executing the plan in `both` mode, and therefore has the appropriate auth cache key metadata attached. We do not attach this metadata to the Rust PlanNode **when in comparison mode**, which results in this false mismatch. When **just the Rust QP is used**, the cache key metadata is present and accounted for correctly, which is why none of the mocks fail when running with only the Rust planner.