do not compare authorization field in PlanNode
Here we skip comparing PlanNode.authorization field when running comparisons between JS and Rust query plans. This field is added _after_ the query plan that will be executed has been created directly in the router, which means it's not going to be different between the two plans. We do not attach this metadata to the Rust PlanNode in comparison mode, since only the JS PlanNode will be executed.