QP comparison mode: skip comparing PlanNode.authorization (#6287)
We are getting mismatches when comparing PlanNodes between JS & Rust with ~authorisation~ authorization directives. This mismatch comes from `authorization` field in PlanNode. The executing query plan (and in comparison mode, it will be the JS QP) receives the appropriate `CacheKeyMetadata`,

```rust
CacheKeyMetadata { 
  is_authenticated: true,
  scopes: [],
  policies: []
}
```
but the Rust QP does not, since it will not be used to execute, authenticate, etc any of the requests and gets discarded in this comparison mode. 

Because we do not attach this metadata to the Rust PlanNode **in comparison mode**, we see a ton of false mismatches for everyone using auth directives. 

This PR skips comparing `authorization` field in PlanNode, as this field gets populated _after_ the query plan is already present. In addition, this data is not based on the query plan itself, so it's not necessary to establish equivalence between the two implementations.

When **just the Rust QP is used**, the cache key metadata is present and accounted for correctly, which is why none of the mocks fail when running with only the Rust planner.

It's incredibly difficult to write a test for this little nit in semantic equivalence, so this PR does not come with any, apologies.
<!-- ROUTER-861 -->