Move `test_updated` to a scheduled workflow and remove from other workflows
Presently, knowing about this condition as soon as possible is valuable in
ensuring that we are quickly aware and able to reconcile changes to
dependencies which may be difficult for library-consumers of the Router.
Today, we get that through a feedback loop on individual PRs and release
workflows, however those can be inconveniently timed and blocking.

By moving this to a scheduled workflow in CircleCI, we can run it at a
cadence that suits us and get notifications via Slack immediately.

This should conceivably supersede https://github.com/apollographql/router/pull/5935