fix(federation): API schema should always include defer definition
Update `to_api_schema()` logic to always add `@defer/@stream` definitions. This addresses the problem that we would error out when user opts-out of `incremental_delivery` feature but pass an operation with `@defer` in it. While the passed in document is valid, during the conversion from `apollo-rs` selections to the federation representations, we explicitly revalidate whether fragment spread directives are valid and present in the API schema. Since API schema may not contain `@defer` definition, this conversion will error out.

Existence of `@defer/@stream` directives in the API schema was driven by the user configuration
We expose a configuration that allows users to enable(default)/disable incremental delivery (`@defer`) support. If incremental delivery is disabled, we normalize the passed in query by removing any and all `@defer` applications. In RS implementation we first convert