fix(federation): filter out invalid fragment spread directives instead of error
Operation was already validated so the passed in document should be `Valid`. Instead of throwing an explicit error if fragment spread dectives are not available in the api schema OR if they are not applicable on an inline fragment, we should be filtering those instead.

This specifically addresses an issue when users disable `incremental_delivery` support (i.e. `@defer`) but pass in a query with `@defer`. Instead of throwing an exception we should be gracefully handling it as all `@defer` usages will be normalized away (i.e. removed) if user opted out of `incremental_delivery` feature.