Fix snapshots after #6205 was merged
`dev` wasn't merge in #6205 before it was squashed and committed.
Because of that `dev` had some changes in tests that didn't change hash produced by old query hash agorithm but resulted in a hash change once #6205 was merged.
Example: #6310 Added test changes related to @context and #6205 added a special case to handle @context directives