fix: use the original unescaped string for single-quoted strings
If the value is not a block string, it may include escaped characters like newlines. We need to use the "unescaped" value because our sers like JSONSelection will puke on doubly-escaped characters like `\\n`. This is to avoid issues if the SDL content is reformatted by a tool that collapses block strings into single-quoted strings.

Offsets will be wrong for non-block strings with escaped racters but manually writing \n will be uncommon (especially in JSONSelection where all whitespace is insignificant).