Reduce cloning of private keys
rustls private keys no longer implement `Clone`. You can instead
explicitly clone them with the `.clone_key()` method. The reason for
that is to discourage keeping lots of copies of private keys in memory.

We can easily avoid cloning of private keys in our configuration module
by wrapping the relevant sections of config in an `Arc`. I'm not sure
this actually reduces copies, since I'm not really aware of when we
clone configuration.

The other place we use `key.clone()` is when setting up the TLS client
configuration for subgraph requests. It's probably possible to avoid
cloning the private keys there, but it's not a security regression to
clone them, so I've used `key.clone_key()` there for the time being.