Fix parameter types for axum handlers
In particular, the body type is now `axum::body::Body`.

Instead of closures for handlers, pass in the options for request
handling via an axum Extension.

The genericity over `RouterFactory` is propagated to `handle_graphql`.

In the future, I think it would be better to use boxed trait objects for
all the factory types, so we can use `#[axum::debug_handler]`.