Fix coprocessor empty body panic If a coprocessor responds with an empty body at the supergraph stage then the router would panic as there was a false expectation that the query had already been validated.
There is a deeper issue around coprocessors and their ability to modify the query before processing. As query analysis and pluginss may use the query to populate context modifying the query won't necessarily do what the user wants. In fact the router will simply ignore the modified query as the parsed version is already in context.