validation for explicit `@key`s
If the author defines a resolvable `@key`, they are declaring that there's an entity resolver for that type.

When creating fetch targets for each connector, we create edge conditions from the variables used in the connector, ignoring the author-defined `@key`s.

Defining `@key`s is helpful to avoid needing `@shareable` (keys are always shareable), and to silence hints about inconsistent keys across subgraphs.

This validation avoids confusion if a key is defined that doesn't match the implementation. Defining a resolvable `@key` requires that the entity resolver is implemented either as:

- An explicit `entity: true` connector
- An "implicit" entity resolver for resolving the non-root parent type of a connected field

```graphql
type T @key(fields: "id") {
  id: ID!
  # this creates an implicit entity resolver so that we can "enter" the subgraph via T to reach T.f
  f: String @connect(http: { GET: "http://x/${this.id}" } selection: "$")
}
```

<!-- https://apollographql.atlassian.net/browse/CNN-275 -->