Use rustls `add_parsable_certificates` to add native certs
This fixes a borrow-checking error, as `roots.add(cert)` takes ownership
of `cert`, and we can't then print it out if adding the cert failed.

The code inside `add_parsable_certificates` is exactly the same as this,
except it uses an internal function that avoids cloning the cert.

I'm not sure why that function was copied into the router source,
perhaps because it did not use to take an iterator?