Remove the dance in next_reponse()
I don't think we need the dance in next_response() because the various
representations of stream, etc are all working on the same data block
and we have exclusive access to it during execution of next_response().

I tested the change manually and it seems fine with defer and normal
queries.