Make changes to address new RUSTSEC details
RUSTSEC-2024-0421 identifies issues in the idna crate at versions <
1.0.3.

We are not affected by the vulnerability, since the router is only
using the crate to resolve subgraph addresses from a trusted source.
We'll ignore the advisory.

When we can update the hickory crate to a version using idna >=1.0.3,
we'll remove the deny override.

Additionally:
    Update the url crate to 2.5.4 anyway, so it's using idna 1.0.3