Use `__typename` object field to track output shapes.
The special `{ <Type> }` syntax for selection sets still works as a
shorthand for `{ __typename: "Type" }`, and may be more useful for
static analysis because it enforces `Type` is an identifier. We may even
want to forbid setting `__typename` to anything other than a (union of)
string literal shapes, or forbid setting `__typename` directly
altogether, in favor of the `<Type>` syntax, since it provides the
flexibility we're comfortable with.