The various aws crates aren't at the same version
There's a comment in the Cargo.toml which implies various AWS crates
need to be at the same version. However, they aren't locked and
Cargo.lock shows that the selected versions vary a lot.

I've added a note and set the (unlocked) versions to match the values
select in Cargo.lock and we can re-visit this at some point.