fix(federation): do not use internal errors for context parse errors
Internal errors should only be used for situations that never happen.
Technically a context parse error should not happen if the supergraph is
valid, so it could be argued to be a correct usage of an internal error.

The test suite contains checks for context parse errors though.
`bail!()` should be treated as a panic, and the tests are not treating
it as a panic, so I think the best way for now is to introduce a new
error branch for `@fromContext` parse errors.