Try to tidy up the axum dependencies a little and tweak hyper
I read through the axum dependencies carefully and it looks like default
includes the ones we wanted, but didn't include "http2", so I changed
the axum entries in Cargo.toml.

I also enabled all of our hyper connections to be "upgrade-able", since
I was concerned that would be a change in behaviour if we didn't do
that. I also wondered if it might help with our "ws" test fails, but it
didn't.