Fix server used for WebSocket tests
This is the cause of the WebSocket test failures:
https://github.com/snapview/tungstenite-rs/pull/363

In short, tungstenite in 0.22.0 started to validate that if you
requested a particular protocol, the server actually declared support
for that protocol.

In tests, our clients request the `graphql-transport-ws` protocol, but
the mock server doesn't declare support for it, so the client now fails
to connect.

This is worth calling out in the 2.0 migration guide, because it's
plausible that a misbehaving subgraph would work in 1.x, but no longer
work in 2.0.