First draft of a backpressure enabled pipeline
There's a long way to go, but some interesting aspects of the required
changes are implemented here.

I noted that a significant number of errors and glitches and tests are
resolved by making these changes. Mainly places where we were calling a
service without readying it first or places where we were cloning an
inner service without doing the memory "replace" dance.

Because we are using a "trick" to make the router service cloneable
right now, there are a few tests which don't work properly. I think for
the "full" work, we'll need to make the router service properly
cloneable (without requiring a mutex).

This will require some fairly substantial re-working of a wide variety
of services and layers. On the plus side, once we've done that work
we'll be able to retire a bunch of code that we've written that we will
no longer require.

I'll pick this up in the New Year...