chore(router): add test for `Accept` with multiple types
Looking at https://github.com/apollographql/router/issues/6501, I didn't
find a test that checks that multiple `Accept` mime types are parsed
correctly. The existing tests either use a single mime type in the
`Accept` header, or use multiple types that are all unsupported by the
router.

The test passes already, so this does not solve #6501, but it covers a
case we had not covered before.