fix: include mutation root type definition in connectors
this fixes a query planning panic when using `@defer` within a mutation.

the root issue is that we never added `schema { mutation: Mutation }` when necessary. you'd think that would be a bigger problem, but apparently the query planner doesn't mind a missing root type definition. if `Mutation` is defined, it'll use it — this is probably a bug!

however, some part of defer bookkeeping cares about the root type definition, so when it's missing, it wasn't able to populate a map of deferred information. this lead to a panic in this very specific case.

in addition, i fixed a TODO about using the router's api schema config within the expansion code.