connectors debugging snapshot test
because the debugger captures the host/port of the http API, i had to add the ability to specify a port. this could be a problem if another test chooses the same port, but that's pretty unlikely given the number of snapshot tests we have.

also, you have to use the subgraph name "connectors" for snapshot tests to work.