Add experimental non-reloadable config for worker pool size
1.58.0 introduced a thread pool for compute-heavy work like GraphQL
validation and query planning: https://github.com/apollographql/router/blob/bd12e44f0e47b0e5e31a75494a6336165c8c91e6/CHANGELOG.md#move-heavy-computation-to-a-thread-pool-with-a-priority-queue-pr-6247

The worker pool uses all available parallelism. Some customers are
reporting concerns that this leads to CPU saturation when a lot of
queries need to be planned, causing other requests to time out.

This branch exists to let select customers experiment with the worker
pool size, and gather data so we can decide on a better default or
introduce useful configuration in a future release.

This introduces a new configuration option:

```yaml
experimental_worker_pool:
  size: 2 # number of threads
```

To avoid significantly refactoring the worker pool implementation,
this configuration is *not* runtime-reloadable, so users must restart
their routers (eg. recycle pods) for changes to take effect.

If we want to release this feature properly, we should do the
refactoring to make the worker pool resizeable at runtime.