Fix the rate limiting test so that it works with new rate limiting
IMPORTANT:

This change modifies the supergraph method invocation test to be a
router_service service invocation test.

Amongst other important details (such as we are now really testing the
service through the full pipeline), it's important to note that we can't
use `oneshot()` and just re-create the service every time we want to
call it. If we do, then the rate limiting details are lost. So, we must
re-use our service to make sure that state isn't lost.