Fix the `apollo_router_session_count_total` metric
In #6495 I migrated this metric to an otel ObservableGauge. However the
lifetime of the count guard was too short, so the value of the metric
would almost always just be 0; and the lifetime of the instrument was
too short, so even the wrong value of 0 would not actually be reported.

Now the lifetime of the instrument is captured by all requests, so the
instrument is kept even post-reload if a request is still being
completed from the old schema.

Also, a test verifies that the count increments and decrements as
expected.